remove session

+++ internal/modules/user/helpers/auth.go
@@ -15,10 +15,11 @@ func Auth(ctx *gin.Context) userResponse.User {
 
 	var response userResponse.User
 
-	authID := sessions.Get(ctx, "auth")
+	authID := sessions.Get(ctx, "auth") 
 	userID, _ := strconv.Atoi(authID)
+	
 	userRepo := userRepository.New()
-	user := userRepo.FindByID(uint(userID))
+	user := userRepo.FindByID(userID)
 
 	if user.ID == 0 {
 		return response
